+++ changelog.md
@@ -246,6 +246,8 @@ Updated:
 - Removed opinium core singularity and replaced it with a AE2 singularity singularity (should be much easier now and more balanced)
 - Changed basic upgrade to -20 (addition) per upgrade to a maximum of 4 per machine
 - Updating custommachinery-1.16.5-0.5.8f causes filters to break do not update
+- Added logic,engineering,calculation processor wafers
+- Added Glassy Redstone Sheet, and Charged Redstone Sheet
 
 ### Mods Added/Updated
 - Updated forge 26.2.35